fix: Extract resources
