+++ app/src/main/java/com/example/letschat/MainActivity.kt
@@ -3,6 +3,8 @@ package com.example.letschat
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
+import androidx.appcompat.app.AppCompatDelegate
+import androidx.appcompat.app.AppCompatDelegate.MODE_NIGHT_NO
 import androidx.databinding.DataBindingUtil
 import com.example.letschat.databinding.ActivityMainBinding
 
@@ -13,5 +15,6 @@ class MainActivity : AppCompatActivity() {
         val binding: ActivityMainBinding = DataBindingUtil.setContentView(
             this, R.layout.activity_main)
         Log.i("Main", "Activity created")
+        AppCompatDelegate.setDefaultNightMode(MODE_NIGHT_NO)
     }
 }
