+++ app/src/main/java/com/example/letschat/repository/LetschatDatabase.kt
@@ -1,9 +1,10 @@
-package com.example.letschat.database
+package com.example.letschat.repository
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.example.letschat.data.Device
 
 @Database(entities = [Device::class], version= 1, exportSchema = false )
 abstract class LetschatDatabase: RoomDatabase(){
