+++ app/src/main/java/com/example/letschat/repository/resources.kt
@@ -1,0 +1,5 @@
+package com.example.letschat.repository
+
+sealed class NetworkResult<out T>{
+
+}
